+++ database/java/migrations/ticket-order-db-migrations/ticket-order-analytical-migrations/src/main/resources/db/migrations/V1.0002__create_events.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS t_event (
+CREATE TABLE IF NOT EXISTS events (
     event_id UUID PRIMARY KEY,
     owner_id UUID NOT NULL,
     date TIMESTAMP WITH TIME ZONE NOT NULL,
@@ -9,11 +9,3 @@ CREATE TABLE IF NOT EXISTS t_event (
     created_at TIMESTAMP WITH TIME ZONE NOT NULL,
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL
 );
-
-ALTER TABLE t_event
-    ADD CONSTRAINT t_event_owner_id_fk
-        FOREIGN KEY (owner_id) REFERENCES t_users(id);
-
-ALTER TABLE t_event
-    ADD CONSTRAINT t_event_status_check
-        CHECK (status IN ('DRAFT', 'PUBLISHED', 'CANCELLED'));