+++ database/java/migrations/ticket-order-db-migrations/ticket-order-analytical-migrations/src/main/resources/db/repeatable/R__users.sql
@@ -1,9 +1,0 @@
-CREATE OR REPLACE VIEW ticket_analytical.users AS
-SELECT
-    id,
-    email,
-    role,
-    enabled,
-    created_at,
-    updated_at
-FROM ticket_transactional.t_users;