+++ services/java/ticket-order-api/src/main/java/com/example/ticketplatform/api/adapter/out/persistence/event/EventEntity.java
@@ -21,7 +21,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "t_event")
+@Table(name = "t_event", schema = "ticket_transactional")
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)